refactor: share update-check script helpers
